moved utils into expression context
